Move builtin-function-related tests to execution_builtin_test.go
